+++ Application/Warcraft3PowerLevels/Entities/Humans/Summons/Phoenix.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Humans.Summons
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Magic;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 68;
+        public double Attack { get; set; } = 68;
         public double AttackTime { get; set; } = 1.4;
         public int Range { get; set; } = 600;
