+++ Application/Warcraft3PowerLevels/Entities/Humans/Summons/WaterElemental_1.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Humans.Summons
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Piercing;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 20;
+        public double Attack { get; set; } = 20;
         public double AttackTime { get; set; } = 1.5;
         public int Range { get; set; } = 300;
