+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Bear_1.cs
@@ -18,7 +18,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Armor { get; set; } = 4;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
 
-        public int Attack { get; set; } = 24;
+        public double Attack { get; set; } = 24;
         public double AttackTime { get; set; } = 1.5;
         public int Range { get; set; } = 100;
