+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Bear_2.cs
@@ -18,7 +18,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Armor { get; set; } = 5;
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Heavy;
 
-        public int Attack { get; set; } = 29;
+        public double Attack { get; set; } = 29;
         public double AttackTime { get; set; } = 1.5;
         public int Range { get; set; } = 100;
