+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Summons/Quilbeast_2.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Summons
         public int Armor { get; set; } = 2;
 
         public ArmorTypeEnum ArmorType { get; set; } = ArmorTypeEnum.Medium;
-        public int Attack { get; set; } = 24;
+        public double Attack { get; set; } = 24;
         public double AttackTime { get; set; } = 1.5;
         public int Range { get; set; } = 100;
