+++ Application/Warcraft3PowerLevels/Entities/Neutrals/Units/HarpyWindwitch.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Neutral.Mercenaries
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Magic;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 24;
+        public double Attack { get; set; } = 24;
         public double AttackTime { get; set; } = 1.7;
         public int Range { get; set; } = 600;
