+++ Application/Warcraft3PowerLevels/Entities/NightElves/Summons/Treant.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Summons
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Normal;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Melee;
-        public int Attack { get; set; } = 16;
+        public double Attack { get; set; } = 16;
         public double AttackTime { get; set; } = 1.75;
         public int Range { get; set; } = 100;
