+++ Application/Warcraft3PowerLevels/Entities/NightElves/Units/GlaiveThrower.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.NightElves.Units
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Siege;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 44;
+        public double Attack { get; set; } = 44;
         public double AttackTime { get; set; } = 3.5;
         public int Range { get; set; } = 1150;
