+++ Application/Warcraft3PowerLevels/Entities/Orcs/Units/WitchDoctor.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Orcs.Units
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Magic;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 12;
+        public double Attack { get; set; } = 12;
         public double AttackTime { get; set; } = 1.75;
         public int Range { get; set; } = 600;
