+++ Application/Warcraft3PowerLevels/Entities/Undead/Units/Destroyer.cs
@@ -14,7 +14,7 @@ namespace Warcraft3PowerLevels.Entities.Undead.Units
 
         public AttackTypeEnum AttackType { get; set; } = AttackTypeEnum.Magic;
         public WeaponTypeEnum WeaponType { get; set; } = WeaponTypeEnum.Ranged;
-        public int Attack { get; set; } = 20;
+        public double Attack { get; set; } = 20;
         public double AttackTime { get; set; } = 1.35;
         public int Range { get; set; } = 500;
